Moved saveAs(tool) to util/saveAs.js, as it just a function call, not a tool.
