+++ index.html
@@ -144,7 +144,7 @@
 		const index = element.getAttribute('data-index');
 		const saveBtn = element.parentElement.querySelector(`a[data-index='${index}']`);
 		saveBtn.addEventListener('click', (evt) => {
-			cornerstoneTools.SaveAsTool(element, 'viewportImage.png')
+			cornerstoneTools.SaveAs(element, 'viewportImage.png')
 		});
 	})
